requirements.txt
PyYaml
conan
pathos
